requirements-vendor.txt
Django==1.9
httplib2==0.9.2
google-api-python-client==1.6.4
requests==2.18.4
